requirements.txt
fastapi
python-jose
passlib
pytest
sqlalchemy
python-dotenv
pydantic
# psycopg2 binary via apt-get
alembic
httpx
python-multipart